+++ examples/example-app-router-minimal/src/app/[locale]/page.tsx
@@ -1,0 +1,6 @@
+import {useTranslations} from 'next-intl';
+
+export default function IndexPage() {
+  const t = useTranslations('IndexPage');
+  return <h1>{t('title')}</h1>;
+}