+++ examples/example-app-router-minimal/src/i18n.ts
@@ -1,0 +1,12 @@
+import {notFound} from 'next/navigation';
+import {getRequestConfig} from 'next-intl/server';
+import {locales} from './navigation';
+
+export default getRequestConfig(async ({locale}) => {
+  // Validate that the incoming `locale` parameter is valid
+  if (!locales.includes(locale as any)) notFound();
+
+  return {
+    messages: (await import(`../messages/${locale}.json`)).default
+  };
+});