+++ examples/example-app-router-playground/src/navigation.tsx
@@ -7,7 +7,8 @@ export const defaultLocale = 'en';
 
 export const locales = ['en', 'de', 'es'] as const;
 
-export const localePrefix = 'as-needed';
+export const localePrefix =
+  process.env.NEXT_PUBLIC_LOCALE_PREFIX === 'never' ? 'never' : 'as-needed';
 
 export const pathnames = {
   '/': '/',