+++ packages/next-intl/src/navigation/shared/BaseLink.tsx
@@ -2,10 +2,17 @@
 
 import NextLink from 'next/link';
 import {usePathname} from 'next/navigation';
-import React, {ComponentProps, forwardRef, useEffect, useState} from 'react';
+import React, {
+  ComponentProps,
+  MouseEvent,
+  forwardRef,
+  useEffect,
+  useState
+} from 'react';
 import useLocale from '../../react-client/useLocale';
 import {LocalePrefix} from '../../shared/types';
 import {isLocalHref, localizeHref, prefixHref} from '../../shared/utils';
+import syncLocaleCookie from './syncLocaleCookie';
 
 type Props = Omit<ComponentProps<typeof NextLink>, 'locale'> & {
   locale: string;
@@ -13,7 +20,7 @@ type Props = Omit<ComponentProps<typeof NextLink>, 'locale'> & {
 };
 
 function BaseLink(
-  {href, locale, localePrefix, prefetch, ...rest}: Props,
+  {href, locale, localePrefix, onClick, prefetch, ...rest}: Props,
   ref: Props['ref']
 ) {
   // The types aren't entirely correct here. Outside of Next.js
@@ -39,6 +46,11 @@ function BaseLink(
       : href
   );
 
+  function onLinkClick(event: MouseEvent<HTMLAnchorElement>) {
+    syncLocaleCookie(pathname, defaultLocale, locale);
+    if (onClick) onClick(event);
+  }
+
   useEffect(() => {
     if (!pathname || localePrefix === 'never') return;
 
@@ -57,7 +69,14 @@ function BaseLink(
   }
 
   return (
-    <NextLink ref={ref} href={localizedHref} prefetch={prefetch} {...rest} />
+    <NextLink
+      ref={ref}
+      href={localizedHref}
+      hrefLang={isChangingLocale ? locale : undefined}
+      onClick={onLinkClick}
+      prefetch={prefetch}
+      {...rest}
+    />
   );
 }